refactor(architecture): move lasts steps to module structure
