+++ antenna-documentation/src/site/markdown/processors/license-knowledgebase-resolver.md
@@ -9,6 +9,6 @@ Add the following step into the `<processors>` section of your workflow.xml
 ```
 <step>
     <name>License Knowledgebase Resolver</name>
-    <classHint>org.eclipse.sw360.antenna.workflow.processors.enricher.LicenseKnowledgeBaseResolver</classHint>
+    <classHint>org.eclipse.sw360.antenna.workflow.processors.LicenseKnowledgeBaseResolver</classHint>
 </step>
 ```
