--- antenna-workflow-steps/processors/validators/antenna-match-state-validator/pom.xml
@@ -1,64 +1,0 @@
-<!--
-  ~ Copyright (c) Bosch Software Innovations GmbH 2016.
-  ~
-  ~ All rights reserved. This program and the accompanying materials
-  ~ are made available under the terms of the Eclipse Public License v2.0
-  ~ which accompanies this distribution, and is available at
-  ~ http://www.eclipse.org/legal/epl-v20.html
-  ~
-  ~ SPDX-License-Identifier: EPL-2.0
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.eclipse.sw360.antenna</groupId>
-		<artifactId>antenna-workflow-steps</artifactId>
-		<version>${revision}${qualifier}</version>
-		<relativePath>../../..</relativePath>
-	</parent>
-
-	<artifactId>antenna-match-state-validator</artifactId>
-
-	<packaging>jar</packaging>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>antenna-model</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>abstract-antenna-compliance-checker</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-api</artifactId>
-			<scope>compile</scope>
-		</dependency>
-		<!-- ################################ testing dependencies ################################ -->
-		<dependency>
-			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>antenna-core-common-testing</artifactId>
-			<version>${project.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.assertj</groupId>
-			<artifactId>assertj-core</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<!-- ################################ compliance dependency ########################### -->
-		<dependency>
-			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>antenna-license-provider</artifactId>
-			<version>1</version>
-			<classifier>licenseinfo</classifier>
-			<type>zip</type>
-			<scope>provided</scope>
-		</dependency>
-	</dependencies>
-</project>
