+++ antenna/core-workflow-steps/src/test/java/org/eclipse/sw360/antenna/workflow/processors/SourceUrlResolverTest.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.sw360.antenna.workflow.processors.enricher;
+package org.eclipse.sw360.antenna.workflow.processors;
 
 import org.eclipse.sw360.antenna.model.artifact.Artifact;
 import org.eclipse.sw360.antenna.model.artifact.facts.ArtifactSourceFile;
