+++ modules/simple-validators/antenna-coordinates-validator/pom.xml
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright (c) Bosch Software Innovations GmbH 2018.
+  ~ Copyright (c) Bosch Software Innovations GmbH 2016.
   ~
   ~ All rights reserved. This program and the accompanying materials
   ~ are made available under the terms of the Eclipse Public License v2.0
@@ -14,16 +14,20 @@
 
 	<parent>
 		<groupId>org.eclipse.sw360.antenna</groupId>
-		<artifactId>antenna-workflow-steps</artifactId>
+		<artifactId>antenna-simple-validators-module</artifactId>
 		<version>${revision}${qualifier}</version>
-		<relativePath>../../..</relativePath>
 	</parent>
 
-	<artifactId>antenna-security-issue-validator</artifactId>
+	<artifactId>antenna-coordinates-validator</artifactId>
 
 	<packaging>jar</packaging>
 
 	<dependencies>
+		<dependency>
+			<groupId>org.eclipse.sw360.antenna</groupId>
+			<artifactId>antenna-model</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sw360.antenna</groupId>
 			<artifactId>antenna-model</artifactId>
