+++ modules/simple-validators/antenna-coordinates-validator/src/main/java/org/eclipse/sw360/antenna/workflow/processors/CoordinatesValidator.java
