+++ modules/simple-validators/antenna-coordinates-validator/src/test/java/org/eclipse/sw360/antenna/workflow/processors/CoordinatesValidatorTest.java
