+++ modules/simple-validators/antenna-license-validator/pom.xml
@@ -14,21 +14,15 @@
 
 	<parent>
 		<groupId>org.eclipse.sw360.antenna</groupId>
-		<artifactId>antenna-workflow-steps</artifactId>
+		<artifactId>antenna-simple-validators-module</artifactId>
 		<version>${revision}${qualifier}</version>
-		<relativePath>../../..</relativePath>
 	</parent>
 
-	<artifactId>antenna-coordinates-validator</artifactId>
+	<artifactId>antenna-license-validator</artifactId>
 
 	<packaging>jar</packaging>
 
 	<dependencies>
-		<dependency>
-			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>antenna-model</artifactId>
-			<version>${project.version}</version>
-		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sw360.antenna</groupId>
 			<artifactId>antenna-model</artifactId>
@@ -65,5 +59,5 @@
 			<type>zip</type>
 			<scope>provided</scope>
 		</dependency>
-	</dependencies>
+    </dependencies>
 </project>
