+++ modules/simple-validators/antenna-license-validator/src/main/java/org/eclipse/sw360/antenna/workflow/processors/LicenseValidator.java
