+++ modules/simple-validators/antenna-license-validator/src/test/java/org/eclipse/sw360/antenna/workflow/processors/LicenseValidatorTest.java
