+++ modules/simple-validators/antenna-security-issue-validator/pom.xml
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright (c) Bosch Software Innovations GmbH 2016.
+  ~ Copyright (c) Bosch Software Innovations GmbH 2018.
   ~
   ~ All rights reserved. This program and the accompanying materials
   ~ are made available under the terms of the Eclipse Public License v2.0
@@ -14,12 +14,11 @@
 
 	<parent>
 		<groupId>org.eclipse.sw360.antenna</groupId>
-		<artifactId>antenna-workflow-steps</artifactId>
+		<artifactId>antenna-simple-validators-module</artifactId>
 		<version>${revision}${qualifier}</version>
-		<relativePath>../../..</relativePath>
 	</parent>
 
-	<artifactId>antenna-license-validator</artifactId>
+	<artifactId>antenna-security-issue-validator</artifactId>
 
 	<packaging>jar</packaging>
 
@@ -60,5 +59,5 @@
 			<type>zip</type>
 			<scope>provided</scope>
 		</dependency>
-    </dependencies>
+	</dependencies>
 </project>
