+++ modules/simple-validators/antenna-security-issue-validator/src/test/java/org/eclipse/sw360/antenna/workflow/processors/SecurityIssueValidatorTest.java
