+++ modules/simple-validators/antenna-source-validator/pom.xml
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright (c) Bosch Software Innovations GmbH 2018.
+  ~ Copyright (c) Bosch Software Innovations GmbH 2016.
   ~
   ~ All rights reserved. This program and the accompanying materials
   ~ are made available under the terms of the Eclipse Public License v2.0
@@ -14,12 +14,11 @@
 
 	<parent>
 		<groupId>org.eclipse.sw360.antenna</groupId>
-		<artifactId>antenna-workflow-steps</artifactId>
+		<artifactId>antenna-simple-validators-module</artifactId>
 		<version>${revision}${qualifier}</version>
-		<relativePath>../../..</relativePath>
 	</parent>
 
-	<artifactId>antenna-source-url-resolver</artifactId>
+	<artifactId>antenna-source-validator</artifactId>
 
 	<packaging>jar</packaging>
 
@@ -31,7 +30,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>antenna-core</artifactId>
+			<artifactId>abstract-antenna-compliance-checker</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
