+++ modules/simple-validators/pom.xml
@@ -1,0 +1,31 @@
+<!--
+  ~ Copyright (c) Bosch Software Innovations GmbH 2016-2017.
+  ~
+  ~ All rights reserved. This program and the accompanying materials
+  ~ are made available under the terms of the Eclipse Public License v2.0
+  ~ which accompanies this distribution, and is available at
+  ~ http://www.eclipse.org/legal/epl-v20.html
+  ~
+  ~ SPDX-License-Identifier: EPL-2.0
+  -->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.eclipse.sw360.antenna</groupId>
+        <artifactId>antenna-modules</artifactId>
+        <version>${revision}${qualifier}</version>
+    </parent>
+
+    <artifactId>antenna-simple-validators-module</artifactId>
+    <name>antenna-simple-validators-module</name>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>antenna-coordinates-validator</module>
+        <module>antenna-license-validator</module>
+        <module>antenna-match-state-validator</module>
+        <module>antenna-security-issue-validator</module>
+        <module>antenna-source-validator</module>
+    </modules>
+</project>
