+++ pom.xml
@@ -49,7 +49,6 @@
         <module>antenna-testing</module>
         <module>antenna-license-provider</module>
         <module>antenna-csv-license-knowledge-base</module>
-        <module>antenna-workflow-steps</module>
         <module>antenna-documentation</module>
     </modules>
